% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miso.R
\name{print.miso_fit}
\alias{print.miso_fit}
\title{Print a fitted MiSo model}
\usage{
\method{print}{miso_fit}(x, ...)
}
\arguments{
\item{x}{A fitted \code{miso_fit} object.}

\item{...}{Additional arguments; currently unused.}
}
\value{
The fitted object invisibly.
}
\description{
Print a fitted MiSo model
}
