% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miso_fit_path.R
\name{print.miso_fit_path}
\alias{print.miso_fit_path}
\title{Print a MiSO fitting path}
\usage{
\method{print}{miso_fit_path}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=miso_fit_path]{miso_fit_path()}} result.}

\item{...}{Additional arguments passed to the summary table's print method.}
}
\value{
The path, invisibly.
}
\description{
Print a MiSO fitting path
}
