% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misoR-package.R
\docType{package}
\name{misoR-package}
\alias{misoR}
\alias{misoR-package}
\title{Mixture of sparse Poisson factor models}
\description{
Fit MiSo using \code{\link[=miso_fit]{miso_fit()}}, create reusable starting parameters with
\code{\link[=miso_init]{miso_init()}}, and fit row-wise Poisson SuSiE with \code{\link[=poisson_susie_fit]{poisson_susie_fit()}}.
Use \code{\link[=miso_fit_length]{miso_fit_length()}} for document-length-adjusted Poisson means.
Observations are rows and features are columns throughout the package.
}
\author{
\strong{Maintainer}: Dat Do \email{dodat.stats@gmail.com}

Authors:
\itemize{
  \item Dat Do \email{dodat.stats@gmail.com}
}

}
\keyword{internal}
